Remove unnecessary interface for ConfigMapper and split out proxy generation into a separate class ProxyFactory.
